requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
elasticsearch==9.3.0
python-dotenv==1.2.1
PyMuPDF==1.27.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
